bot/requirements.txt
python-telegram-bot==20.3
psycopg2-binary
requests
google-cloud-documentai
google-api-core
redis>=5.0.1
asyncio